// src/models/pss-message.interface.ts
export interface PSSMessage {
  MessageId: string;
  UserId: string;
  Message: string;
  UserName: string;
  UserSpriteId: string;
  MessageType: string;
  ChannelId: string;
  ActivityType: string;
  ActivityArgument: string;
  Argument: string;
  AllianceName: string;
  AllianceSpriteId: string;
  AllianceId: string;
  MessageDate: string;
  Trophy: string;
  ShipDesignId: string;
  BorderSpriteId: string;
}